connected

+++ frontend/src/component/State/Restaurant/Action.js
@@ -71,7 +71,7 @@ export const getRestaurantById = (reqData) => {
 export const getRestaurantsByUserId = (token) => async (dispatch) => {
     dispatch({ type: GET_RESTAURANT_BY_USER_ID_REQUEST });
     try {
-        const res = await api.get("/api/restaurants/user");
+        const res = await api.get("/api/admin/restaurants/user");
         dispatch({ type: GET_RESTAURANT_BY_USER_ID_SUCCESS, payload: res.data });
     } catch (error) {
         console.log("Error in fetching restaurant by user ID data", error)
@@ -82,7 +82,7 @@ export const getRestaurantsByUserId = (token) => async (dispatch) => {
 export const createRestaurant = (reqData) => async (dispatch) => {
     dispatch({ type: CREATE_RESTAURANT_REQUEST });
     try {
-        const res = await api.post("/api/restaurants", reqData);
+        const res = await api.post("/api/admin/restaurants", reqData);
         dispatch({ type: CREATE_RESTAURANT_SUCCESS, payload: res.data });
     } catch (error) {
         dispatch({ type: CREATE_RESTAURANT_FAILURE, payload: error.response.data });
@@ -92,7 +92,7 @@ export const createRestaurant = (reqData) => async (dispatch) => {
 export const updateRestaurant = (reqData) => async (dispatch) => {
     dispatch({ type: UPDATE_RESTAURANT_REQUEST });
     try {
-        const res = await api.put(`/api/restaurants/${reqData.restaurantId}`, reqData);
+        const res = await api.put(`/api/admin/restaurants/${reqData.restaurantId}`, reqData);
         dispatch({ type: UPDATE_RESTAURANT_SUCCESS, payload: res.data });
     } catch (error) {
         console.log("Error in updating restaurant data", error)
@@ -103,7 +103,7 @@ export const updateRestaurant = (reqData) => async (dispatch) => {
 export const deleteRestaurant = (restaurantId) => async (dispatch) => {
     dispatch({ type: DELETE_RESTAURANT_REQUEST });
     try {
-        const res = await api.delete(`/api/restaurants/${restaurantId}`);
+        const res = await api.delete(`/api/admin/restaurants/${restaurantId}`);
         dispatch({ type: DELETE_RESTAURANT_SUCCESS, payload: res.data });
     } catch (error) {
         dispatch({ type: DELETE_RESTAURANT_FAILURE, payload: error.response.data });
@@ -113,7 +113,7 @@ export const deleteRestaurant = (restaurantId) => async (dispatch) => {
 export const updateRestaurantStatus = (reqData) => async (dispatch) => {
     dispatch({ type: UPDATE_RESTAURANT_STATUS_REQUEST });
     try {
-        const res = await api.put(`/api/restaurants/${reqData.restaurantId}/status`, reqData);
+        const res = await api.put(`/api/admin/restaurants/${reqData.restaurantId}/status`, reqData);
         dispatch({ type: UPDATE_RESTAURANT_STATUS_SUCCESS, payload: res.data });
     } catch (error) {
         console.log("Error in updating restaurant status data", error)
@@ -165,7 +165,7 @@ export const getRestaurantEvents = (restaurantId, token) => async (dispatch) => 
     }
 }
 
-export const createCategoryAction = (reqData) => async (dispatch) => {
+export const createCategoryAction = ({reqData, jwt}) => async (dispatch) => {
     dispatch({ type: CREATE_CATEGORY_REQUEST });
     try {
         const res = await api.post("/api/admin/category", reqData);
@@ -179,7 +179,7 @@ export const createCategoryAction = (reqData) => async (dispatch) => {
 export const getRestaurantCategory = ({jwt, restaurantId}) => async (dispatch) => {
     dispatch({ type: GET_RESTAURANT_CATEGORY_REQUEST });
     try {
-        const res = await api.get(`/api/category/restaurant/${restaurantId}`);
+        const res = await api.get(`/api/admin/category/category/restaurant`);
         dispatch({ type: GET_RESTAURANT_CATEGORY_SUCCESS, payload: res.data });
     } catch (error) {
         console.log("Error in fetching restaurant category data", error)
